Support PKCS#8 keys
Instead of selecting the keys by reading the header, we do it through
more generic OpenSSL functions.

This may not be convenient or the way expected as it may not be
compatible with other libraries, but it works for now.

In case we go forward with this, we should also cleanup other functions.