Session callback when an exception is received
When the socket receives an error, a set of callbacks are internally handling
all the process of changing the status of the connection, but nothing is
communicated to the outside. This requires the loops to continuously check if
the session is still connected, but this is not practical in a fully
asynchronous scenario.