Dispatch reading too much data causes too long packet nested loop
The packet reading loop on libssh is nesting ssh_packet_socket_callback
functions, and when the received length is big, it may cause a pile of thousands
of functions. This will cause the runtime to break with a BAD_ACCESS code 2.

When reading data for dispatch, it is not necessary to loop as the CFStream will
throw the event again if not all data has been read. This will guarantee that we
do not read too much and potentially break.

This was causing random reads whenever there was too much data waiting in the
channel.